+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Adds `Range.prototype.isEmpty`.
 - Adds `Range.prototype.intersects`.
+- Adds `Range.prototype.contains`.
 
 ## 1.0.0 (May 11, 2015)
 - Adds support for setting range bounds with `new Range(1, 3, "[)")` to set